requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.17.0
charset-normalizer==3.0.1
cryptography==43.0.0
Django==4.2.15
django-allauth==64.1.0
django-prometheus==2.3.1
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
django-smtp-ssl==1.0
django-utils-six==2.0
gunicorn==20.1.0
idna==3.4
oauthlib==3.2.2
pillow==11.1.0
prometheus-client==0.17.1
psycopg2-binary==2.9.5
pycparser==2.22
PyJWT==2.9.0
requests==2.28.2
requests-oauthlib==2.0.0
sqlparse==0.4.3
urllib3==1.26.14
